\name{gatingMLData-package}
\alias{gatingMLData-package}
\alias{gatingMLData}
\docType{package}

\title{
Gating-ML test data and XML files
}

\description{
Test data and XML files for testing compliance of the 
flowUtils/flowCore packages with Gating-ML (1.5 and 2.0) standards.
The data is used by the flowUtils package - it is
required for the testGatingMLCompliance method.
}

\details{
\tabular{ll}{
Package: \tab gatingMLData\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
J. Spidlen, N. Gopalakrishnan

Maintainer: J. Spidlen <jspidlen@bccrc.ca>
}

\keyword{ package }

