\name{gatingMLData-package}
\alias{gatingMLData-package}
\alias{gatingMLData}
\docType{package}

\title{
GatingML test Data and XML files
}

\description{
Test data files for testing compliance with the
gating-ML V1.5 candidate recommendataion. The data is
used by the flowUtils package .
}

\details{
\tabular{ll}{
Package: \tab gatingMLData\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{
Nishant Gopalakrishnan

Maintainer:  Nishant Gopalakrishnan <ngopalak@fhcrc.org>
}

\keyword{ package }

